+++ site-cookbooks/apache_part/spec/recipes/default_spec.rb
@@ -1,19 +1,13 @@
 require_relative '../spec_helper'
 
 describe 'apache_part::default' do
-  let(:chef_run) do
-    ChefSpec::SoloRunner.new(
-      cookbook_path: %w(cookbooks site-cookbooks),
-      platform: 'centos',
-      version: '6.5'
-    ).converge(described_recipe)
-  end
+  let(:chef_run) { ChefSpec::SoloRunner.converge(described_recipe) }
 
   before do
     stub_command('/usr/sbin/httpd -t').and_return(0)
   end
 
-  it 'install apache_part::setup' do
+  it 'include setup recipe' do
     expect(chef_run).to include_recipe('apache_part::setup')
   end
 end