+++ site-cookbooks/postgresql_part/spec/recipes/default_spec.rb
@@ -1,16 +1,10 @@
 require_relative '../spec_helper'
-require 'chefspec'
 
 describe 'postgresql_part::default' do
-  let(:chef_run) do
-    ChefSpec::SoloRunner.new(
-      cookbook_path: %w(cookbooks site-cookbooks),
-      platform: 'centos',
-      version: '6.5'
-    ).converge(described_recipe)
-  end
+  let(:chef_run) { ChefSpec::SoloRunner.converge(described_recipe) }
 
-  it 'include postgresql_part::setup' do
+  it 'include setup recipe' do
+    stub_command('ls /var/lib/pgsql/9.3/data/recovery.conf')
     expect(chef_run).to include_recipe 'postgresql_part::setup'
   end
 end