+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+## version 0.3.2 (2014/12/24)
+
+  - Support latest serverspec.
+  - Add default data directory for postgresql.
+  - Add default CIDR to CloudConductorLocation parameter.
+  - Add dependencies between SubnetRouteTableAssociation and EIP to specify remove order.
+  - Brush up chef recipes on apache_part and postgresql_part.
+
 ## version 0.3.1 (2014/11/17)
 
   - Modify default_attribute for opscode cookbooks changed in postgresql_part